requirements.txt
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
psutil==5.7.3
pycodestyle==2.6.0
python-gnupg==0.4.6
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
